vim-patch:8.1.0395: compiler warning on 64-bit MS-Windows
Problem:    Compiler warning on 64-bit MS-Windows.
Solution:   Add type cast. (Mike Williams)

https://github.com/vim/vim/commit/6e272acc82af900318017061f923e7f66dc7ee7a

+++ src/nvim/diff.c
@@ -692,7 +692,7 @@ static int diff_write_buffer(buf_T *buf, diffin_T *din)
 
   // xdiff requires one big block of memory with all the text.
   for (lnum = 1; lnum <= buf->b_ml.ml_line_count; lnum++) {
-    len += STRLEN(ml_get_buf(buf, lnum, false)) + 1;
+    len += (long)STRLEN(ml_get_buf(buf, lnum, false)) + 1;
   }
   ptr = xmalloc(len);
   if (ptr == NULL) {
